# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Game Review API",
    "description": "Collects and Displays Video Game Review Information",
    "version": "1.0.0"
  },
  "host": "week5-8-30xe.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/login": {
      "get": {
        "description": "serves a login page",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "Passes the login information for verification",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "test@test.com"
                },
                "password": {
                  "example": "The user's password"
                }}}}],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/reviews/": {
      "get": {
        "description": "gets all the reviews in the database",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Add a new Review",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "author": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "musicR": {
                  "example": "any"
                },
                "gameplayR": {
                  "example": "any"
                },
                "storyR": {
                  "example": "any"
                },
                "overallRating": {
                  "example": "any"
                },
                "comments": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "418": {
            "description": "I'm a teapot"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/games/": {
      "get": {
        "description": "Games in the database",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "genre": {
                  "example": "any"
                },
                "platform": {
                  "example": "any"
                },
                "releaseDate": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}